requirements.txt
pandas
numpy
seaborn
matplotlib
scikit-learn
torch
transformers
groq>=0.4.0
python-dotenv